feat: Add support for querying chart data by user ID
The code changes include modifications to the backend `ChartsController` and `ChartsService` files to support querying chart data by user ID. The `getChartData` method in the `ChartsController` now accepts a `userId` query parameter, which is used to retrieve the chart data specific to the given user. The `ChartsService` has been updated to handle this new parameter and extract the user ID from the JWT. Additionally, the frontend `ChartCard` component has been modified to pass the user ID to the `useFetchChartData` hook.

Recent user commits:
- (fix) chart conditions, rm dead code
- Refactor chart condition filtering logic in EditChartPanel component
- (save) progress
- (feat) add, remove conditions, wip
- (save) progress

Recent repository commits:
- (fix) chart conditions, rm dead code
- Refactor chart condition filtering logic in EditChartPanel component
- (save) progress
- (feat) add, remove conditions, wip
- (save) progress

```